FIX: Relizamos los cambios sugeridos
